Split out the proxy and DB process, and re-write most of proxy.
Message passing happens between the proxy and the db writer via
protobufs and they communicate over a socket.  Things are still
a little flakey due to the sheer quantity of data being written
over the socket (the entire JSON request basically).  This needs
to be fixed to create individual protobufs for interesting events
and only writing the data we're interested in.