Make a centralized GameState.
This way there is a centralized location for all known information
about the current state of the game.  As a result of this, the
active battles tab page can now show data for a battle that was
entered before the user ever clicked on the tab.

Previously it relied on getting events through an event handler
that was set up in the form's OnLoad event, but that event is not
called until the first time the page is shown.  This way, the
OnLoad handler can just look at the GameState class to initialize
the list views.

Also, we now handle all the events related to terminating a battle
encounter (win, loss, escape) and correctly update the game state
and UI accordingly.