Refactor the way responses are processed.
Instead of having a giant if/else statement, moved the logic into
smaller more isolated ResponseHandler classes which do all the
processing.  A nice side benefit of this is that it makes lazy
parsing of the json (i.e. only parse if this request is going to
be handled) very clean and straightforward, which is nice since
we want to spend as little time as possible processing each
response (lest the server think it has lost its connection to us).