Make the "All Drops for this Dungeon" list view work.
This includes an overhaul of the way database requests are
processed.  In particular, we now have a request queue because
otherwise doesn't know to serialize requests to the database
thread and it breaks when you try to issue a request when another
is still outstanding.

To support this, all of the actual database operations (e.g. code
that manipulates MySqlCommand objects, etc) has been moved to
individual classes that implement a single database request.

This also allows for much cleaner and smaller implementations,
since all of the transaction code can be abstracted out to a
higher level and reused across all requests.

One of the requirements for getting this all workign was to update
the schema to change the dungeon_drops view to contain more info
about the enemies.  So that is done as well, as adding a foreign
key constraint from drops to enemies.

Finally, we now record enemy information about the end of each
encounter.