Fix data caching to work again, and implement the "All Drops" list.
Now on the Current Battle tab, the All Drops for this dungeon list
view works correctly and updates after each battle, or when the
program first launches.  As a result of having data caching, we also
can now display the item name and other information for the active
battle round as well.