Make schema version 6.  Adds world series columns.
The series is necessary on the dungeons table as well as the
world table, and all the procedures / views that make use of
those tables.  We were just using the realm (I, II, IX, X, etc)
before, which is usually sufficient, but it turns out that this
isn't always true.  There are some series ids that have special
meanings, so we have to store actual series ids.

This also fixes up the client side code appropriately to refer
to new column names, etc.