Update missing / wrong item information on /dff/party/list.
Now when you view the party screen and you posess an item that
the server does not know about, it will update the database.
If the item was previously missing, it will add it to the items
table, and if the item existed but had an incorrect name (usually
because the database was initially populated without the realm name
suffix) it will fix the name.

Eventually this should be removed, but for now it makes populating
the database and fixing mistakes easier.