Add schema 14.
The standard deviation related statistics are wiped from the
database, and instead we now store statistics as a histogram.
That is, for each combination of (battle,item) we now also
store a histogram bucket and a histogram value.  Since the range
of possible values for number of drops is small, this is a very
efficient way to store the full history of drops without needing
to store every value.  And it is lossless, so we can use a variety
of different algorithms to analyze the data after the fact.

The UI columns for variance and margin of error are removed
temporarily, until I have a chance to put them back in using
the new histogram mechanic.