Fix an error in handling of list battles request.
If this was the first instance of anyone ever viewing the battles
for a particular dungeon, we would try to insert some entries
into the cache.  But there was a bug in this path that was causing
a null pointer dereference.

This only happened it was the first time EVER that someone had
entered a dungeon.  So the impact on data wasn't huge, but a few
records may have been lost if you ran one of these dungeons many
times.