Add a basic item stats editor for inputting item stats.
This is a debug-only feature, and the editing tab is not even
added to the control unless your database is connected to
localhost.

also this adds schema 17, which massages the items table a little
bit and fixes some inconsistencies with the data, as well as
adding an equipment_stats table which will hold the basics necessary
to compute synergized item stats and analyze inventory.