Some improvement to the item stats editor.
It's now always enabled, but editing of existing items' stats
is hidden behind a #define.  If the #define is not enabled, you
get only the option to add missing items, not editing existing
ones.  Adding missing items writes to a different table that is
not the main items table, so it gives the chance for the changes
to be reviewed before comitting to the main database.