Fixed handling of Ctrl+Right and Ctrl+Left.
The editing control swallows these which leads to moving the
selected cell to the end or beginning of a row if you use
Ctrl+Right or Ctrl+Left while editing.

This seems like a bug in DataGridView, but anyway the fix is to
subclass DataGridView and handle the key ourselves.