Dump of database after purging bad data
Unfortunately due to an error I made some time ago, some data had been
recorded in a way that was straight up wrong.  It wasn't correlated with
time, but rather the popularity of the dungeon, and it was mostly
unpopular dungeons which were affected, although a few popular ones were
affected.  I was able to identify exactly which Battles were affected,
and only reset those, but unfortunately a rollback is a rollback, and
it's never pretty.  At least i found this sooner rather than later, so
life can continue on.