Don't permanently kill connectivity if the connection gets lost.
We have a background through pumping messages from the database
request queue.  If the connection to the database server gets lost
for any reason, then an exception would get thrown all the way out
of the database thread function, effectively killing connectivity
to the database for the duration of the Fiddler session.  The only
way to restore connectivity is to shutdown and restart Fiddler.

With this patch, if the connection is lost it will always try to
re-establish.