Read character equip usage from /dff/party/list.
I was relying on the database for this information, but it turns
out this is not necessary, as the game sends us all this info.
This is great, as it means that eventually we can delete the
characters and equip_usage tables entirely.  For now, all the UI
for managing equipment usage is gone, and this also makes it so
we never have to update the database for new characters when they
come out.